Move SNBTReader/Writer from nbt.io to nbt.io.snbt
